unzip and install files

+++ readsy/MasterViewController.m
@@ -8,6 +8,7 @@
 
 #import "MasterViewController.h"
 #import "DetailViewController.h"
+#import "DataFileViewController.h"
 #import "ReadsyMetadata.h"
 #import "AppDelegate.h"
 #import "Constants.h"
@@ -51,6 +52,27 @@
 - (void)viewDidAppear:(BOOL)animated
 {
     [[self tableView] reloadData];
+    
+    // check for pending file uploads
+    NSUserDefaults *defaults = [NSUserDefaults standardUserDefaults];
+    if ([defaults objectForKey:kDataUploadInProgress]) {
+        UIAlertController *alert = [UIAlertController alertControllerWithTitle:@"Continue?"
+                                                                       message:@"There is a data file upload in progress. Would you like to continue uploading files now?"
+                                                                preferredStyle:UIAlertControllerStyleAlert];
+        UIAlertAction *yes = [UIAlertAction actionWithTitle:@"Yes"
+                                                      style:UIAlertActionStyleDefault
+                                                    handler:^(UIAlertAction *action) {
+                                                        [self performSegueWithIdentifier:@"handleDataFile" sender:nil];
+                                                    }];
+        UIAlertAction *no = [UIAlertAction actionWithTitle:@"No"
+                                                     style:UIAlertActionStyleCancel
+                                                   handler:nil];
+        [alert addAction:yes];
+        [alert addAction:no];
+        [self.navigationController presentViewController:alert
+                                                animated:YES
+                                              completion:nil];
+    }
 }
 
 - (void)initDropbox
@@ -78,12 +100,7 @@
             [self.objects removeAllObjects];
             [[self tableView] reloadData];
         }
-        UIAlertView *alert = [[UIAlertView alloc] initWithTitle:@"Dropbox Not Linked"
-                                                        message:@"There is no Dropbox account linked with readsy. To link your Dropbox account, tap Settings."
-                                                       delegate:nil
-                                              cancelButtonTitle:@"OK"
-                                              otherButtonTitles: nil];
-        [alert show];
+        [self showDropboxNotLinkedAlert];
     }
     
 }
@@ -106,13 +123,28 @@
         if (self.restClient) {
             self.restClient = nil;
         }
-        UIAlertView *alert = [[UIAlertView alloc] initWithTitle:@"Dropbox Not Linked"
-                                                        message:@"There is no Dropbox account linked with readsy. To link your Dropbox account, tap Settings."
-                                                       delegate:nil
-                                              cancelButtonTitle:@"OK"
-                                              otherButtonTitles: nil];
-        [alert show];
+        [self showDropboxNotLinkedAlert];
     }
+}
+
+-(void)showDropboxNotLinkedAlert {
+    UIAlertController *alert = [UIAlertController alertControllerWithTitle:@"Dropbox Not Linked"
+                                                                   message:@"There is no Dropbox account linked with readsy. To link your Dropbox account, tap Settings. Would you like to go to settings now?"
+                                                            preferredStyle:UIAlertControllerStyleAlert];
+    UIAlertAction *yes = [UIAlertAction actionWithTitle:@"Yes"
+                                                  style:UIAlertActionStyleDefault
+                                                handler:^(UIAlertAction *action) {
+                                                    [self performSegueWithIdentifier:@"showSettingsSegue" sender:nil];
+                                                }];
+    UIAlertAction *no = [UIAlertAction actionWithTitle:@"No"
+                                                 style:UIAlertActionStyleCancel
+                                               handler:nil];
+    [alert addAction:yes];
+    [alert addAction:no];
+    
+    [self.navigationController presentViewController:alert
+                                            animated:YES
+                                          completion:nil];
 }
 
 - (void)viewWillDisappear:(BOOL)animated
@@ -130,13 +162,6 @@
     NSLog(@"**********************MEMORY WARNING");
 }
 
-#pragma mark - UIAlertViewDelegate
-- (void)alertView:(UIAlertView *)alertView clickedButtonAtIndex:(NSInteger)buttonIndex
-{
-    if (buttonIndex == 1) {
-        [[UIApplication sharedApplication] openURL:[NSURL URLWithString:ReadsyHelpURL]];
-    }
-}
 
 
 #pragma mark - Dropbox Access
@@ -152,23 +177,37 @@
             [array addObject:file.filename];
         }
         if (array.count == 0) {
-            UIAlertView *alert = [[UIAlertView alloc] initWithTitle:@"Nothing To Read"
-                                                            message:@"It looks like you do not have any data files in Dropbox. To learn more about how to install and create data files, visit the readsy website."
-                                                           delegate:self
-                                                  cancelButtonTitle:@"Not Now"
-                                                  otherButtonTitles:@"Visit Website", nil];
-            [alert show];
+            UIAlertController *alert = [UIAlertController alertControllerWithTitle:@"Nothing To Read"
+                                                                           message:@"It looks like you do not have any data files in Dropbox. To learn more about how to install and create data files, visit the readsy website."
+                                                                    preferredStyle:UIAlertControllerStyleAlert];
+            UIAlertAction *yes = [UIAlertAction actionWithTitle:@"Visit Website"
+                                                          style:UIAlertActionStyleDefault
+                                                        handler:^(UIAlertAction *action) {
+                                                            [[UIApplication sharedApplication] openURL:[NSURL URLWithString:ReadsyHelpURL]];
+                                                        }];
+            UIAlertAction *no = [UIAlertAction actionWithTitle:@"Not Now"
+                                                         style:UIAlertActionStyleCancel
+                                                       handler:nil];
+            [alert addAction:yes];
+            [alert addAction:no];
+            [self.navigationController presentViewController:alert
+                                                    animated:YES
+                                                  completion:nil];
         } else {
             NSArray *sortedArray = [array sortedArrayUsingSelector:@selector(caseInsensitiveCompare:)];
             
             for (NSString *file in sortedArray) {
-                ReadsyMetadata *rm = [[ReadsyMetadata alloc] initWithSourceDirectory:file];
-                [self.objects addObject:rm];
-                [self.tableView reloadData];
-                NSString *tmpFile = [NSString stringWithFormat:@"%@%@", NSTemporaryDirectory(), file];
-                [self showActivityIndicators:YES];
-                //            [AppDelegate setActivityIndicatorsVisible:YES];
-                [self.restClient loadFile:[NSString stringWithFormat:@"/%@/metadata", file] intoPath:tmpFile];
+                if ([file hasSuffix:@"_tmp_"]) {
+                    NSLog(@"Skipping incomplete upload '%@'", file);
+                } else {
+                    ReadsyMetadata *rm = [[ReadsyMetadata alloc] initWithSourceDirectory:file];
+                    [self.objects addObject:rm];
+                    [self.tableView reloadData];
+                    NSString *tmpFile = [NSString stringWithFormat:@"%@%@", NSTemporaryDirectory(), file];
+                    [self showActivityIndicators:YES];
+                    //            [AppDelegate setActivityIndicatorsVisible:YES];
+                    [self.restClient loadFile:[NSString stringWithFormat:@"/%@/metadata", file] intoPath:tmpFile];
+                }
             }
         }
     }
@@ -177,7 +216,7 @@
 - (void)restClient:(DBRestClient *)client loadMetadataFailedWithError:(NSError *)error
 {
     [self showActivityIndicators:NO];
-    [self showErrorMessage];
+    [self showErrorMessage:nil];
     NSLog(@"Error loading metadata: %@", error);
 }
 
@@ -187,8 +226,7 @@
     NSError *error;
     NSString *readsyMetadata = [NSString stringWithContentsOfFile:localPath encoding:NSUTF8StringEncoding error:&error];
     if (error) {
-        NSLog(@"There was an error reading the file - %@", error);
-        [self showErrorMessage];
+        [self showErrorMessage:[NSString stringWithFormat:@"Unable to read file '%@'.", localPath]];
     } else {
         for (ReadsyMetadata *rm in self.objects) {
             if ([rm.sourceDirectory isEqualToString:[localPath lastPathComponent]]) {
@@ -205,18 +243,64 @@
 
 - (void)restClient:(DBRestClient*)client loadFileFailedWithError:(NSError*)error {
     [self showActivityIndicators:NO];
-    NSLog(@"There was an error loading the file - %@", error);
-    [self showErrorMessage];
+    NSString *message = nil;
+    // if we cannot load metadata, we look at the error message
+    // and determine the path of the file we were loading.
+    //
+    // Then, we try to remove that ReadsyMetadata object from the data model,
+    // and refresh the table view so that it doesn't show a stuck "Loading" message
+    // for the failed load.
+    //
+    // The path dictionary entry looks like this: "path=/Test/metadata" -- we need path component number 1, not 0,
+    // because 0 would be the leading "/"
+    NSString *sourceDirFailed = [[[[error userInfo] objectForKey:@"path"] pathComponents] objectAtIndex:1];
+    if (sourceDirFailed) {
+        message = [NSString stringWithFormat:@"Unable to load file '%@' from Dropbox. This may be due to a network error, or to missing files on Dropbox.", sourceDirFailed];
+        for (ReadsyMetadata *rm in self.objects) {
+            if ([rm.sourceDirectory isEqualToString:sourceDirFailed]) {
+                [self.objects removeObject:rm];
+            }
+            [self.tableView reloadData];
+        }
+    }
+    
+    [self showErrorMessage:message];
 }
 
-- (void)showErrorMessage
-{
-    UIAlertView *alert = [[UIAlertView alloc] initWithTitle:@"Error"
-                                                    message:@"There was an error while communicating with Dropbox. There may be a network problem."
-                                                   delegate:nil
-                                          cancelButtonTitle:@"OK"
-                                          otherButtonTitles:nil];
-    [alert show];
+- (void)restClient:(DBRestClient *)client deletedPath:(NSString *)path {
+    NSLog(@"Delete path '%@' successful", path);
+}
+
+- (void)restClient:(DBRestClient *)client deletePathFailedWithError:(NSError *)error {
+    UIAlertController *alert = [UIAlertController alertControllerWithTitle:@"Delete Failed"
+                                                                   message:@"Unable to delete the data from Dropbox at this time. Try again later."
+                                                            preferredStyle:UIAlertControllerStyleAlert];
+    UIAlertAction *ok = [UIAlertAction actionWithTitle:@"OK"
+                                                 style:UIAlertActionStyleDefault
+                                               handler:^(UIAlertAction *action) {
+                                                   [self refresh];
+                                               }];
+    [alert addAction:ok];
+    [self presentViewController:alert
+                       animated:YES
+                     completion:nil];
+}
+
+- (void)showErrorMessage: (NSString *)message {
+    if (!message) {
+        message = @"There was an error while communicating with Dropbox. There may be a network problem.";
+    }
+    
+    UIAlertController *alert = [UIAlertController alertControllerWithTitle:@"Error"
+                                                                   message:message
+                                                            preferredStyle:UIAlertControllerStyleAlert];
+    UIAlertAction *ok = [UIAlertAction actionWithTitle:@"OK"
+                                                 style:UIAlertActionStyleDefault
+                                               handler:nil];
+    [alert addAction:ok];
+    [self.navigationController presentViewController:alert
+                                            animated:YES
+                                          completion:nil];
 }
 
 -(BOOL)splitViewController:(UISplitViewController *)svc shouldHideViewController:(UIViewController *)vc inOrientation:(UIInterfaceOrientation)orientation
@@ -282,6 +366,35 @@
     }
 }
 
+- (void)tableView:(UITableView *)tableView commitEditingStyle:(UITableViewCellEditingStyle)editingStyle forRowAtIndexPath:(NSIndexPath *)indexPath
+{
+    if (editingStyle == UITableViewCellEditingStyleDelete) {
+        ReadsyMetadata *rm = [self.objects objectAtIndex:indexPath.row];
+        
+        UIAlertController *alert = [UIAlertController alertControllerWithTitle:@"Delete?"
+                                                                       message:[NSString stringWithFormat:@"You are about to delete %@. This will remove the data from Dropbox. Are you sure?", rm.fileDescription]
+                                                                preferredStyle:UIAlertControllerStyleAlert];
+        UIAlertAction *ok = [UIAlertAction actionWithTitle:@"OK"
+                                                     style:UIAlertActionStyleDestructive
+                                                   handler:^(UIAlertAction *action) {
+                                                       [self.restClient deletePath:[NSString stringWithFormat:@"/%@", rm.sourceDirectory]];
+                                                       [self.objects removeObjectAtIndex:indexPath.row];
+                                                       [tableView deleteRowsAtIndexPaths:[NSArray arrayWithObject:indexPath]
+                                                                        withRowAnimation:UITableViewRowAnimationFade];
+                                                   }];
+        UIAlertAction *cancel = [UIAlertAction actionWithTitle:@"Cancel"
+                                                         style:UIAlertActionStyleDefault
+                                                       handler:^(UIAlertAction *action) {
+                                                           [tableView setEditing:NO animated:YES];
+                                                       }];
+        
+        [alert addAction:ok];
+        [alert addAction:cancel];
+        
+        [self presentViewController:alert animated:YES completion:nil];
+    }
+}
+
 - (void)prepareForSegue:(UIStoryboardSegue *)segue sender:(id)sender
 {
     if ([[segue identifier] isEqualToString:@"showDetail"]) {
@@ -289,6 +402,9 @@
         ReadsyMetadata *object = [self.objects objectAtIndex:indexPath.row];
         [[segue destinationViewController] setDetailItem:object];
         [[segue destinationViewController] setTitle:object.fileShortDescription];
+    } else if ([[segue identifier] isEqualToString:@"handleDataFile"]) {
+        [[segue destinationViewController] setDataFileDelegate:self];
+        [[segue destinationViewController] setDataFile:sender];
     }
 }
 
@@ -323,6 +439,19 @@
         [self.refreshControl endRefreshing];
         [[UIApplication sharedApplication] setNetworkActivityIndicatorVisible:NO];
         self.navigationItem.title = @"Library";
+    }
+}
+
+#pragma mark - Import data file methods
+- (void)handleDataFile:(NSURL *)dataFile
+{
+    NSLog(@"Handling data file...");
+    [self performSegueWithIdentifier:@"handleDataFile" sender:dataFile];
+}
+
+- (void)dataLoadFinished:(BOOL)finished {
+    if (finished) {
+        [self.objects removeAllObjects];
     }
 }
